Restructuring Data
Change the structure for the database, currently setupTables.py is not complete and will not interact with any tables in db.

added csv files to represent the new db structure.

started to rewrite setupTables.py to fit the new structure.

TODO:

finish rewriting setupTables.py
edit readCSV.py so one can get a df for any of the csv files in the directory
edit analysis code to accomodate for new structure where needed.