Progressed with updateTables.py
Split tables into two categories: childs and not childs. Childs have foreign key constraints
and non childs do not. Decided not to call non-childs parents as some parents in the db are
also children.

Wrote a function to get the id for a new entry from the non-child tables and update those tables
if necessary.

Wrote a function to update the dates table which is both a parent and child. Also wrote one to update
this table

TODO:

- write function to get ids from static tables (i.e ones that will never be updated like dismissals
and formats)

- write functions to update fixtures and stats tables.

- write single function to auto update the db when the file is run.

- test functionality to ensure updateTables.py works properly