Completed and wrote tests for updateRecords.py
Wrote the tests and ran for four different test cases as written on file doc string

ran tests so had to fix some bugs in updateRecords.py but all runs smoothly now.

However, currently one table would be updated even if the tables which depends on it may have an error.

Ideally would refactor the code so all tables are updated only when all entries are valid and no errors are thrown.

TODO:

Write more methods of analysis of the data